# user-system-develop/app/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: User System API
  description: |
    Интерактивная документация для проверки HTTP API (пути совпадают с сервером Gin).
    Для защищённых методов нажмите **Authorize** и введите `Bearer <access_token>`.
  version: 1.0.0
servers:
  - url: /
    description: Текущий хост (Try it out)
...
paths:
...
  /api/rpc:
    post:
      tags: [RPC]
      summary: JSON-RPC 2.0
      operationId: jsonRpc
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JsonRpcRequest"
            examples:
              companyHierarchy:
                summary: Иерархия компании
                value:
                  jsonrpc: "2.0"
                  method: company.getHierarchy
                  params:
                    company_id: "00000000-0000-0000-0000-000000000000"
                  id: 1
              userByRole:
                summary: Пользователи по роли
                value:
                  jsonrpc: "2.0"
                  method: user.getByRole
                  params:
                    role_id: "00000000-0000-0000-0000-000000000000"
                  id: 2
      responses:
        "200":
          description: Ответ JSON-RPC
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Вставьте только токен или строку вида `Bearer ...` — Swagger добавит префикс сам.
...
  schemas:
...
    JsonRpcRequest:
      type: object
      required: [jsonrpc, method, id]
      properties:
        jsonrpc:
          type: string
          enum: ["2.0"]
        method:
          type: string
        params:
          type: object
          additionalProperties: true
        id:
          oneOf:
            - type: string
            - type: integer